+++ src/game/client/components/effects.cpp
@@ -80,7 +80,7 @@ void CEffects::SmokeTrail(vec2 Pos, vec2 Vel)
 	p.m_LifeSpan = 0.5f + frandom()*0.5f;
 	p.m_StartSize = 12.0f + frandom()*8;
 	p.m_EndSize = 0;
-	p.m_Friction = 0.7;
+	p.m_Friction = 0.7f;
 	p.m_Gravity = frandom()*-500.0f;
 	m_pClient->m_pParticles->Add(CParticles::GROUP_PROJECTILE_TRAIL, &p);
 }
@@ -185,8 +185,8 @@ void CEffects::Explosion(vec2 Pos)
 			if(x == 0 && y == 0)
 				continue;
 			
-			float a = 1 - (length(vec2(x,y)) / length(vec2(8,8)));
-			m_pClient->m_pFlow->Add(Pos+vec2(x,y)*16, normalize(vec2(x,y))*5000.0f*a, 10.0f);
+			float a = 1 - (length(vec2((float)x,(float)y)) / length(vec2(8,8)));
+			m_pClient->m_pFlow->Add(Pos+vec2((float)x,(float)y)*16, normalize(vec2((float)x,(float)y))*5000.0f*a, 10.0f);
 		}
 		
 	// add the explosion
